+++ app/models/supplier/framework/lot/rate.rb
@@ -6,15 +6,16 @@ class Supplier < ApplicationRecord
         belongs_to :position, inverse_of: :supplier_framework_lot_rates
         belongs_to :jurisdiction, inverse_of: :rates, class_name: 'Supplier::Framework::Lot::Jurisdiction', foreign_key: :supplier_framework_lot_jurisdiction_id
 
-        def rate_in_pounds
+        delegate :rate_type, to: :position
+        delegate :mandatory, to: :position
+
+        def normalized_rate
           rate / 100.0
         end
 
         def rate_as_percentage_decimal
           rate / 10000.0
         end
-
-        alias_method :rate_as_percentage, :rate_in_pounds
       end
     end
   end