+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Update how the show pages are displayed to help with editing ([PR 2159](https://github.com/Crown-Commercial-Service/crown-marketplace-legacy/pull/2159))
+- Add `rate_type` and `mandatory` flag to the positions to improve data model ([PR 2170](https://github.com/Crown-Commercial-Service/crown-marketplace-legacy/pull/2170))
 
 ### Added